requirements-data-collection.txt
# Data collection requirements (includes ML dependencies)
# Install these ONLY when running comment_analysis.py to fetch and process data
# Usage: pip install -r requirements-data-collection.txt

# Core runtime dependencies
pandas
plotly
python_dateutil
streamlit

# Data collection specific dependencies
google_api_python_client
numpy
python-dotenv
scikit_learn
sentence_transformers
tqdm
